% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jdb.R
\name{remove_axis_ticks}
\alias{remove_axis_ticks}
\title{remove ticks and numbers from axes}
\usage{
remove_axis_ticks()
}
\value{
Theme element which removes axis elements
}
\description{
remove ticks and numbers from axes
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point() +
  theme_jdb() +
  remove_axis_ticks()

}
