% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jdb.R
\name{scale_color_pub}
\alias{scale_color_pub}
\title{ggplot2 wong color}
\usage{
scale_color_pub(...)
}
\arguments{
\item{...}{parameters to be passed to \code{ggplot2::discrete_scale()}}
}
\value{
Wong color scale
}
\description{
ggplot2 wong color
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
  geom_point() +
  theme_pub() +
  scale_color_pub()

}
