% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jdb.R
\name{scale_fill_pub}
\alias{scale_fill_pub}
\title{ggplot2 wong color fill}
\usage{
scale_fill_pub(...)
}
\arguments{
\item{...}{parameters to be passed to \code{ggplot2::discrete_scale()}}
}
\value{
Wong color scale (fill)
}
\description{
ggplot2 wong color fill
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, fill = Species)) +
  geom_point(pch = 21) +
  theme_pub() +
  scale_fill_pub()

}
