% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jdb.R
\name{theme_jdb}
\alias{theme_jdb}
\title{ggplot2 theme}
\usage{
theme_jdb(gridlines = FALSE, text = "sans")
}
\arguments{
\item{gridlines}{Should grid lines be included in the plot}
}
\value{
Theme elements to make the style I want.
}
\description{
ggplot2 theme
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point() +
  theme_jdb()

}
