% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jdb.R
\name{theme_pub}
\alias{theme_pub}
\title{ggplot2 publication theme}
\usage{
theme_pub(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\value{
Theme elements to make a "publication style"
}
\description{
ggplot2 publication theme
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point() +
  theme_pub()

}
